+++ entry/Cargo.toml
@@ -32,5 +32,8 @@ solana-logger = { path = "../logger", version = "=1.9.0" }
 crate-type = ["lib"]
 name = "solana_entry"
 
+[[bench]]
+name = "entry_sigverify"
+
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
