+++ entry/src/lib.rs
@@ -1,5 +1,4 @@
 #![allow(clippy::integer_arithmetic)]
-#![feature(test)]
 pub mod entry;
 pub mod poh;
